+++ CODEOWNERS
@@ -52,7 +52,7 @@
 /src/workbench/extensions/manager/                @viva-jinyi @christian-byrne @ltdrdata
 
 # Model-to-node mappings (cloud team)
-/src/stores/modelNodeMappings.ts                  @deepme987
+/src/platform/assets/mappings/                    @deepme987
 
 # LLM Instructions (blank on purpose)
 .claude/
