+++ src/platform/assets/mappings/modelNodeMappings.ts
